Comments!

+++ src/app/simulation/field.rs
@@ -4,11 +4,12 @@ use std::ops::{Index, IndexMut};
 
 use crate::app::simulation::variables::DIVISIONS;
 
+// represents a 1D vector field
 pub struct Field {
-    field: Array1<f32>,
-    extent: Rangef,
-    step: f32,
-    points: Array1<f32>,
+    field: Array1<f32>,  // values of field at each point
+    extent: Rangef,      // dimensions of field
+    step: f32,           // distance between divisions
+    points: Array1<f32>, // x coordinates of each division
 }
 
 impl Field {
@@ -22,6 +23,7 @@ impl Field {
         f
     }
 
+    // get fractional index of value at this x coordinate
     fn index_of(&self, x: f32) -> f32 {
         (x - self.extent.min) / self.step
     }
@@ -34,38 +36,49 @@ impl Field {
         self.field.slice_mut(s![..]).into_slice().unwrap()
     }
 
+    // return interpolated value of field at a given x coordinate
     pub fn value_at(&self, x: f32) -> f32 {
-        let get_value = |i: f32| -> f32 { *self.values().get(i as usize).unwrap_or(&0.0) };
+        // closure returns value of field at a given index, or 0 if index is out of bounds
+        let get_value = |i: usize| -> f32 { *self.values().get(i).unwrap_or(&0.0) };
+        // fractional index of coordinate
         let idx = self.index_of(x);
-        //return *field.get(idx.round() as usize).unwrap_or(&0.0);
+        // closest true indices to the fractional index
         let (lower_idx, upper_idx) = (idx.floor(), idx.ceil());
-        let (lower, upper) = (get_value(lower_idx), get_value(upper_idx));
+        // values at the nearest indices
+        let (lower, upper) = (get_value(lower_idx as usize), get_value(upper_idx as usize));
+        // linearly interpolate to get most accurate field value
         lower * (1.0 - idx + lower_idx) + upper * (idx - lower_idx)
     }
 
+    // given a function of x coordinate and time, fill this field with values at time t
     pub fn set_from_function(&mut self, f: impl Fn(f32, f32) -> f32, t: f32) {
         for i in 0..DIVISIONS {
             self.field[i] = f(self.points[i], t);
         }
     }
 
+    // x coordinates of field divisions
     pub fn intervals(&self) -> &[f32] {
         self.points.slice(s![..]).to_slice().unwrap()
     }
 
+    // add another field to this field
     pub fn add(&mut self, rhs: &Field) {
         self.field += &rhs.field;
     }
 
+    // x coordinate of a division index
     pub fn position_at(&self, idx: usize) -> f32 {
         self.points[idx]
     }
 
+    // dimensions in simulation space
     pub fn size(&self) -> f32 {
         self.extent.span()
     }
 }
 
+// required for [i] operations
 impl Index<usize> for Field {
     type Output = f32;
 
@@ -73,7 +86,6 @@ impl Index<usize> for Field {
         &self.values()[index]
     }
 }
-
 impl IndexMut<usize> for Field {
     fn index_mut(&mut self, index: usize) -> &mut Self::Output {
         &mut self.values_mut()[index]
